+++ lib/hanami/settings.rb
@@ -48,8 +48,8 @@ module Hanami
   #
   # 1. {Hanami::Settings::EnvStore}, which fetches the value from the equivalent upper-cased key in
   #    `ENV`.
-  # 2. {Hanami::Settings::YamlFileStore}, for `config/settings/[HANAMI_ENV].yml`.
-  # 3. {Hanami::Settings::YamlFileStore}, for `config/settings/default.yml`.
+  # 2. {Hanami::Settings::YAMLFileStore}, for `config/settings/[HANAMI_ENV].yml`.
+  # 3. {Hanami::Settings::YAMLFileStore}, for `config/settings/default.yml`.
   #
   # The YAML files are optional, and are read from the app root only: like `ENV`, they are shared by
   # the app and all its slices.
@@ -61,7 +61,7 @@ module Hanami
   # [dry-t]: https://dry-rb.org/gems/dry-types/
   #
   # @see Hanami::Settings::EnvStore
-  # @see Hanami::Settings::YamlFileStore
+  # @see Hanami::Settings::YAMLFileStore
   #
   # @api public
   # @since 2.0.0